api/pyproject.toml
[tool.poetry]
name = "base-api"
version = "0.1.0"
description = "Base API project."
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
psycopg2 = "^2.9.3"
black = "^22.3.0"
isort = "^5.10.1"
pre-commit-hooks = "^4.2.0"
pre-commit = "^2.19.0"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
validators = "^0.19.0"
django-simple-history = "^3.1.1"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
flake8 = "^4.0.1"
PyJWT = "^2.4.0"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
py_version = 38
known_django = ["django"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["venv", ".venv"]
